move j2v8 to sdktest module
