+++ blockstack-sdk/src/androidTest/java/org/blockstack/android/sdk/BlockstackSessionStorageTest.kt
@@ -5,16 +5,12 @@ import android.graphics.drawable.BitmapDrawable
 import android.util.Log
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.ActivityTestRule
-import org.blockstack.android.sdk.model.CryptoOptions
-import org.blockstack.android.sdk.model.GetFileOptions
-import org.blockstack.android.sdk.model.PutFileOptions
-import org.blockstack.android.sdk.model.DeleteFileOptions
-import org.blockstack.android.sdk.model.toBlockstackConfig
+import kotlinx.coroutines.runBlocking
+import org.blockstack.android.sdk.model.*
 import org.blockstack.android.sdk.test.TestActivity
 import org.hamcrest.CoreMatchers.*
 import org.hamcrest.MatcherAssert.assertThat
 import org.json.JSONObject
-import org.junit.After
 import org.junit.Assert.assertTrue
 import org.junit.Assert.fail
 import org.junit.Before
@@ -36,6 +32,7 @@ private val BITCOIN_ADDRESS = "1NZNxhoxobqwsNvTb16pdeiqvFvce3Yg8U"
 
 @RunWith(AndroidJUnit4::class)
 class BlockstackSessionStorageTest {
+    private lateinit var blockstack: Blockstack
     @get:Rule
     val rule = ActivityTestRule(TestActivity::class.java)
 
@@ -43,39 +40,38 @@ class BlockstackSessionStorageTest {
 
     @Before
     fun setup() {
-        session = BlockstackSession(rule.activity,
-                "https://flamboyant-darwin-d11c17.netlify.com".toBlockstackConfig(emptyArray()),
+        blockstack = Blockstack()
+        session = BlockstackSession(
+                appConfig = "https://flamboyant-darwin-d11c17.netlify.com".toBlockstackConfig(emptyArray()),
                 sessionStore = sessionStoreforIntegrationTests(rule),
-                executor = IntegrationTestExecutor(rule))
+                blockstack = blockstack)
     }
 
-    @After
-    fun teardown() {
-        session.release()
-    }
 
     @Test
     fun testPutStringFileTwice() {
         var result1: String? = null
         var result2: String? = null
         val latch = CountDownLatch(2)
+        runBlocking {
 
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", "Hello Test", PutFileOptions(false)) {
-                if (it.value is String) {
-                    result1 = it.value as String
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", "Hello Test", PutFileOptions(false)) {
+                    if (it.value is String) {
+                        result1 = it.value as String
+                    }
+                    latch.countDown()
                 }
+                session.putFile("try.txt", "Hello Test2", PutFileOptions(false)) {
+                    if (it.value != null) {
+                        result2 = it.value as String
+                    }
+                    latch.countDown()
+                }
+            } else {
+                latch.countDown()
                 latch.countDown()
             }
-            session.putFile("try.txt", "Hello Test2", PutFileOptions(false)) {
-                if (it.value != null) {
-                    result2 = it.value as String
-                }
-                latch.countDown()
-            }
-        } else {
-            latch.countDown()
-            latch.countDown()
         }
         latch.await()
         assertThat(result1, `is`(notNullValue()))
@@ -87,9 +83,9 @@ class BlockstackSessionStorageTest {
         assertTrue(session.isUserSignedIn())
         val options = CryptoOptions()
         val plainContent = "hello from test"
-        val encResult = session.encryptContent(plainContent, options = options)
+        val encResult = blockstack.encryptContent(plainContent, options = options)
         assertTrue(encResult.hasValue)
-        val decResult = session.decryptContent(encResult.value!!.json.toString(), false, options)
+        val decResult = blockstack.decryptContent(encResult.value!!.json.toString(), false, options)
         assertThat(decResult.value as String, `is`(plainContent))
     }
 
@@ -98,9 +94,9 @@ class BlockstackSessionStorageTest {
         assertTrue(session.isUserSignedIn())
         val binaryContent = getImageBytes()
         val options = CryptoOptions()
-        val encResult = session.encryptContent(binaryContent, options = options)
+        val encResult = blockstack.encryptContent(binaryContent, options = options)
         assertTrue(encResult.hasValue)
-        val decResult = session.decryptContent(encResult.value!!.json.toString(), true, options)
+        val decResult = blockstack.decryptContent(encResult.value!!.json.toString(), true, options)
         assertThat((decResult.value as ByteArray).size, `is`(binaryContent.size))
     }
 
@@ -109,13 +105,15 @@ class BlockstackSessionStorageTest {
         var result: Result<Any>? = null
         val latch = CountDownLatch(1)
 
-        if (session.isUserSignedIn()) {
-            session.getFile("404file.txt", GetFileOptions(false)) {
-                result = it
+        runBlocking {
+            if (session.isUserSignedIn()) {
+                session.getFile("404file.txt", GetFileOptions(false)) {
+                    result = it
+                    latch.countDown()
+                }
+            } else {
                 latch.countDown()
             }
-        } else {
-            latch.countDown()
         }
         latch.await()
         assertThat(result, `is`(notNullValue()))
@@ -128,17 +126,21 @@ class BlockstackSessionStorageTest {
         var result: String? = null
         val latch = CountDownLatch(1)
 
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", "Hello Test", PutFileOptions(false)) {
-                session.getFile("try.txt", GetFileOptions(false)) {
-                    if (it.value is String) {
-                        result = it.value as String
+        runBlocking {
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", "Hello Test", PutFileOptions(false)) {
+                    runBlocking {
+                        session.getFile("try.txt", GetFileOptions(false)) {
+                            if (it.value is String) {
+                                result = it.value as String
+                            }
+                            latch.countDown()
+                        }
                     }
-                    latch.countDown()
                 }
+            } else {
+                latch.countDown()
             }
-        } else {
-            latch.countDown()
         }
         latch.await()
         assertThat(result, `is`("Hello Test"))
@@ -149,15 +151,18 @@ class BlockstackSessionStorageTest {
         var result: Any? = null
         val latch = CountDownLatch(1)
 
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", "Hello Test", PutFileOptions(false, "application/x.foo")) {
-                val u = URL(it.value).openConnection()
-                u.connect()
-                result = u.contentType
+        runBlocking {
+
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", "Hello Test", PutFileOptions(false, "application/x.foo")) {
+                    val u = URL(it.value).openConnection()
+                    u.connect()
+                    result = u.contentType
+                    latch.countDown()
+                }
+            } else {
                 latch.countDown()
             }
-        } else {
-            latch.countDown()
         }
         latch.await()
         assertThat(result as String, `is`("application/x.foo"))
@@ -168,17 +173,23 @@ class BlockstackSessionStorageTest {
         var result: String? = null
         val latch = CountDownLatch(1)
 
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", "Hello Test", PutFileOptions(true)) {
-                session.getFile("try.txt", GetFileOptions(true)) {
-                    if (it.value is String) {
-                        result = it.value as String
+        runBlocking {
+
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", "Hello Test", PutFileOptions(true)) {
+                    runBlocking {
+
+                        session.getFile("try.txt", GetFileOptions(true)) {
+                            if (it.value is String) {
+                                result = it.value as String
+                            }
+                            latch.countDown()
+                        }
                     }
-                    latch.countDown()
                 }
+            } else {
+                latch.countDown()
             }
-        } else {
-            latch.countDown()
         }
         latch.await()
         assertThat(result, `is`("Hello Test"))
@@ -192,17 +203,23 @@ class BlockstackSessionStorageTest {
 
         val latch = CountDownLatch(1)
 
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", bitMapData, PutFileOptions(false)) {
-                session.getFile("try.txt", GetFileOptions(false)) {
-                    if (it.value is ByteArray) {
-                        result = it.value as ByteArray
+        runBlocking {
+
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", bitMapData, PutFileOptions(false)) {
+                    runBlocking {
+
+                        session.getFile("try.txt", GetFileOptions(false)) {
+                            if (it.value is ByteArray) {
+                                result = it.value as ByteArray
+                            }
+                            latch.countDown()
+                        }
                     }
-                    latch.countDown()
                 }
+            } else {
+                latch.countDown()
             }
-        } else {
-            latch.countDown()
         }
         latch.await()
         assertThat(result?.size, `is`(bitMapData.size))
@@ -216,17 +233,23 @@ class BlockstackSessionStorageTest {
 
         val latch = CountDownLatch(1)
 
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", bitMapData, PutFileOptions(true)) {
-                session.getFile("try.txt", GetFileOptions(true)) {
-                    if (it.value is ByteArray) {
-                        result = it.value as ByteArray
+        runBlocking {
+
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", bitMapData, PutFileOptions(true)) {
+                    runBlocking {
+
+                        session.getFile("try.txt", GetFileOptions(true)) {
+                            if (it.value is ByteArray) {
+                                result = it.value as ByteArray
+                            }
+                            latch.countDown()
+                        }
                     }
-                    latch.countDown()
                 }
+            } else {
+                latch.countDown()
             }
-        } else {
-            latch.countDown()
         }
         latch.await()
         assertThat(result?.size, `is`(bitMapData.size))
@@ -236,19 +259,24 @@ class BlockstackSessionStorageTest {
     fun testPutGetFileSigned() {
         var result: String? = null
         val latch = CountDownLatch(1)
+        runBlocking {
 
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", "Hello Test", PutFileOptions(false, sign = true), {
-                session.getFile("try.txt", GetFileOptions(false, verify = true)) {
-                    if (!it.hasErrors) {
-                        Log.d("blockstack test", it.value as String)
-                        result = it.value as String
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", "Hello Test", PutFileOptions(false, sign = true)) {
+                    runBlocking {
+
+                        session.getFile("try.txt", GetFileOptions(false, verify = true)) {
+                            if (!it.hasErrors) {
+                                Log.d("blockstack test", it.value as String)
+                                result = it.value as String
+                            }
+                            latch.countDown()
+                        }
                     }
-                    latch.countDown()
                 }
-            })
-        } else {
-            latch.countDown()
+            } else {
+                latch.countDown()
+            }
         }
         latch.await()
         assertThat(result, `is`("Hello Test"))
@@ -259,27 +287,34 @@ class BlockstackSessionStorageTest {
         var result: String? = null
         val latch = CountDownLatch(1)
 
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", "Hello Test", PutFileOptions(false, sign = true), {
-                session.deleteFile("try.txt.sig", DeleteFileOptions()) {
-                    if (!it.hasErrors) {
-                        session.getFile("try.txt", GetFileOptions(false, verify = true)) {
+        runBlocking {
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", "Hello Test", PutFileOptions(false, sign = true)) {
+                    runBlocking {
+                        session.deleteFile("try.txt.sig", DeleteFileOptions()) {
                             if (!it.hasErrors) {
-                                result = it.value as String
-                            } else {
-                                result = it.error?.message
-                            }
-                            latch.countDown()
-                        }
-                    } else {
-                        result = "error while deleting signature"
-                        latch.countDown()
-                    }
-                }
+                                runBlocking {
+                                    session.getFile("try.txt", GetFileOptions(false, verify = true)) {
 
-            })
-        } else {
-            latch.countDown()
+                                        if (!it.hasErrors) {
+                                            result = it.value as String
+                                        } else {
+                                            result = it.error?.message
+                                        }
+                                        latch.countDown()
+                                    }
+                                }
+                            } else {
+                                result = "error while deleting signature"
+                                latch.countDown()
+                            }
+                        }
+                    }
+
+                }
+            } else {
+                latch.countDown()
+            }
         }
         latch.await()
         assertThat(result, startsWith("SignatureVerificationError: Failed to verify signature: Failed to obtain signature for file: try.txt"))
@@ -290,17 +325,22 @@ class BlockstackSessionStorageTest {
         var result: JSONObject? = null
         val latch = CountDownLatch(1)
 
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", "Hello Test", PutFileOptions(true, sign = true), {
-                session.getFile("try.txt", GetFileOptions(false)) {
-                    if (!it.hasErrors) {
-                        result = JSONObject(it.value as String)
+        runBlocking {
+
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", "Hello Test", PutFileOptions(true, sign = true)) {
+                    runBlocking {
+                        session.getFile("try.txt", GetFileOptions(false)) {
+                            if (!it.hasErrors) {
+                                result = JSONObject(it.value as String)
+                            }
+                            latch.countDown()
+                        }
                     }
-                    latch.countDown()
                 }
-            })
-        } else {
-            latch.countDown()
+            } else {
+                latch.countDown()
+            }
         }
         latch.await()
         assertThat(result, `is`(notNullValue()))
@@ -314,14 +354,18 @@ class BlockstackSessionStorageTest {
 
         val invalidSignedEncryptedText = "{\"signature\":\"INVALID_SIGNATURE\",\"publicKey\":\"024634ee1d4ff57f2e0ec7a847e1705ec562949f84a83d1f5fdb5956220a9775e0\",\"cipherText\":\"{\\\"iv\\\":\\\"329acaeffe36e8ae58365b56b31af640\\\",\\\"ephemeralPK\\\":\\\"0333fde58c40196efa696dde93fb615e8e960bf52d78ab883d67fb56d4b9a10c5a\\\",\\\"cipherText\\\":\\\"143df68fd1542b29febe2b0843e723af\\\",\\\"mac\\\":\\\"68c3e439c26a2be400aeb278ed7061a8802b0366bf79a1d64a7a6e10e4710047\\\",\\\"wasString\\\":true}\"}"
         if (session.isUserSignedIn()) {
-            session.putFile("try.txt", invalidSignedEncryptedText, PutFileOptions(false), {
-                session.getFile("try.txt", GetFileOptions(true, verify = true)) {
-                    if (it.hasErrors) {
-                        result = it.error?.message
+            runBlocking {
+                session.putFile("try.txt", invalidSignedEncryptedText, PutFileOptions(false)) {
+                    runBlocking {
+                        session.getFile("try.txt", GetFileOptions(true, verify = true)) {
+                            if (it.hasErrors) {
+                                result = it.error?.message
+                            }
+                            latch.countDown()
+                        }
                     }
-                    latch.countDown()
                 }
-            })
+            }
         } else {
             latch.countDown()
         }
@@ -332,77 +376,55 @@ class BlockstackSessionStorageTest {
 
     @Test
     fun getUserAppFileUrlReturns_NO_URL_forNonPublicFile() {
-        val latch = CountDownLatch(1)
-        var url: String? = null
-
-        session.getUserAppFileUrl("non_public_file.txt", "friedger.id", "https://blockstack-todos.appartisan.com/", null) {
-            url = it.value
-            latch.countDown()
-        }
-
-        latch.await()
+        val url = blockstack.getUserAppFileUrl("non_public_file.txt", "friedger.id", "https://blockstack-todos.appartisan.com/", null)
         assertThat(url, `is`("NO_URL"))
     }
 
     @Test
     fun getAppBucketUrlReturnsUrl() {
-        val latch = CountDownLatch(1)
-        var url: String? = null
-
-        session.getAppBucketUrl("https://hub.blockstack.org", PRIVATE_KEY) {
-            url = it.value
-            latch.countDown()
+        val url = runBlocking {
+            blockstack.getAppBucketUrl("https://hub.blockstack.org", PRIVATE_KEY)
         }
 
-        latch.await()
         assertThat(url, `is`("https://gaia.blockstack.org/hub/1NZNxhoxobqwsNvTb16pdeiqvFvce3Yg8U/"))
     }
 
     @Test
     fun listFilesReturnsCorrectNumberOfFiles() {
-        val latch = CountDownLatch(1)
         var count = 0
-        var countResult: Result<Int>? = null
-        session.listFiles({ fileResult ->
-            assertTrue(fileResult.hasValue)
-            count++
-            true
-        }, {
-            countResult = it
-            latch.countDown()
-        })
-        latch.await(1, TimeUnit.MINUTES)
-        assertThat(countResult?.error, `is`(nullValue()))
-        assertThat(countResult?.value, `is`(count))
+        val countResult = runBlocking {
+            session.listFiles {
+                count++
+                true
+            }
+        }
+        assertThat(countResult.value, `is`(count))
     }
 
     @Test
     fun listFilesCanHandleErrorInCallback() {
-        val latch = CountDownLatch(1)
-        var countResult: Result<Int>? = null
-        session.listFiles({ _ ->
-            throw RuntimeException("I want to make the API crash!")
-        }, {
-            countResult = it
-            latch.countDown()
-        })
-        latch.await(1, TimeUnit.MINUTES)
-        assertThat(countResult?.error?.message, `is`("I want to make the API crash!"))
+        val countResult = runBlocking {
+            session.listFiles { _ ->
+                throw RuntimeException("I want to make the API crash!")
+            }
+        }
+        assertThat(countResult.error?.message, `is`("I want to make the API crash!"))
     }
 
     @Test
     fun testGetFileUrlForEncryptedFile() {
         val latch = CountDownLatch(1)
         var urlResult: Result<String>? = null
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", "Hello Test", PutFileOptions(true)) {
-                session.getFileUrl("try.txt", GetFileOptions(true)) {
-                    urlResult = it
-                    latch.countDown()
+        runBlocking {
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", "Hello Test", PutFileOptions(true)) {
+                    runBlocking {
+                        urlResult = session.getFileUrl("try.txt", GetFileOptions(true))
+                    }
                 }
+            } else {
+                latch.countDown()
             }
-        } else {
-            latch.countDown()
         }
         latch.await(1, TimeUnit.MINUTES)
         assertThat(urlResult?.value, notNullValue())
@@ -413,15 +435,16 @@ class BlockstackSessionStorageTest {
     fun testGetFileUrlForUnencryptedFile() {
         val latch = CountDownLatch(1)
         var urlResult: Result<String>? = null
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", "Hello Test", PutFileOptions(false)) {
-                session.getFileUrl("try.txt", GetFileOptions(false)) {
-                    urlResult = it
-                    latch.countDown()
+        runBlocking {
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", "Hello Test", PutFileOptions(false)) {
+                    urlResult = runBlocking {
+                        session.getFileUrl("try.txt", GetFileOptions(false))
+                    }
                 }
+            } else {
+                latch.countDown()
             }
-        } else {
-            latch.countDown()
         }
         latch.await(1, TimeUnit.MINUTES)
         assertThat(urlResult?.value, notNullValue())
@@ -431,15 +454,12 @@ class BlockstackSessionStorageTest {
     @Test
     fun testGetFileUrlFor404File() {
         val latch = CountDownLatch(1)
-        var urlResult: Result<String>? = null
-        session.getFileUrl("404file.txt", GetFileOptions(false)) {
-            urlResult = it
-            latch.countDown()
+        val urlResult = runBlocking {
+            session.getFileUrl("404file.txt", GetFileOptions(false))
         }
-        latch.await(1, TimeUnit.MINUTES)
-        assertThat(urlResult?.value, notNullValue())
+        assertThat(urlResult.value, notNullValue())
         try {
-            URL(urlResult!!.value).readText()
+            URL(urlResult.value).readText()
             fail("Should throw FileNotFoundException")
         } catch (e: FileNotFoundException) {
             // success
@@ -451,17 +471,23 @@ class BlockstackSessionStorageTest {
         var result: Result<Any>? = null
         val latch = CountDownLatch(1)
 
-        if (session.isUserSignedIn()) {
-            session.putFile("try.txt", "Hello Test", PutFileOptions(false)) {
-                session.deleteFile("try.txt") {
-                    session.getFile("try.txt", GetFileOptions(false)) {
-                        result = it
-                        latch.countDown()
+        runBlocking {
+            if (session.isUserSignedIn()) {
+                session.putFile("try.txt", "Hello Test", PutFileOptions(false)) {
+                    runBlocking {
+                        session.deleteFile("try.txt") {
+                            runBlocking {
+                                session.getFile("try.txt", GetFileOptions(false)) {
+                                    result = it
+                                    latch.countDown()
+                                }
+                            }
+                        }
                     }
                 }
+            } else {
+                latch.countDown()
             }
-        } else {
-            latch.countDown()
         }
         latch.await()
         assertThat(result, `is`(notNullValue()))
