+++ blockstack-sdk/src/androidTest/java/org/blockstack/android/sdk/BlockstackSessionTokenTest.kt
@@ -1,8 +1,9 @@
-package org.blockstack.android.sdk;
+package org.blockstack.android.sdk
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.ActivityTestRule
-import com.eclipsesource.v8.V8ScriptExecutionException
+import kotlinx.coroutines.runBlocking
+import me.uport.sdk.jwt.JWTTools
 import org.blockstack.android.sdk.model.Entity
 import org.blockstack.android.sdk.model.Profile
 import org.blockstack.android.sdk.model.toBlockstackConfig
@@ -29,93 +30,89 @@ private val TOKEN_PUBLIC_KEY_2 = "0354a4b971100da0c0be348fc62a30757d115cb4044adf
 private val TOKEN_PROFILE_CONTENT = "{\"@context\":\"http:\\/\\/schema.org\\/\",\"@type\":\"CreativeWork\",\"name\":\"Balloon Dog\",\"creator\":[{\"@type\":\"Person\",\"@id\":\"therealjeffkoons.id\",\"name\":\"Jeff Koons\"}],\"dateCreated\":\"1994-05-09T00:00:00-0400\",\"datePublished\":\"2015-12-10T14:44:26-0500\"}"
 private val DECODED_TOKEN = "{\"header\":{\"typ\":\"JWT\",\"alg\":\"ES256K\"},\"payload\":{\"jti\":\"a35837f5-c8cf-4ba3-b98e-7969a53fec48\",\"iat\":\"2017-03-04T16:13:06.790Z\",\"exp\":\"2018-03-04T16:13:06.790Z\",\"subject\":{\"publicKey\":\"03e7a4d7983369d33ed13020885985d68f8b05dee6168f75f9d97e1a072cdfcdc5\"},\"issuer\":{\"publicKey\":\"03e7a4d7983369d33ed13020885985d68f8b05dee6168f75f9d97e1a072cdfcdc5\"},\"claim\":{\"@context\":\"http:\\/\\/schema.org\\/\",\"@type\":\"CreativeWork\",\"name\":\"Balloon Dog\",\"creator\":[{\"@type\":\"Person\",\"@id\":\"therealjeffkoons.id\",\"name\":\"Jeff Koons\"}],\"dateCreated\":\"1994-05-09T00:00:00-0400\",\"datePublished\":\"2015-12-10T14:44:26-0500\"}},\"signature\":\"MF7ru91rk8IIKNEEqo9wjLHkvW3jSlcDJmeZZeOVSj9KlXApBp67q_3ke0-LzSO_YyYsUnGOplMYiNxY1XynAA\"}"
 private val ZONE_FILE_CONTENT = "\$ORIGIN friedger.id\n\$TTL 3600\n_http._tcp URI 10 1 \"https://gaia.blockstack.org/hub/1Maw8BjWgj6MWrBCfupqQuWANthMhefb2v/0/profile.json\"\n"
-
+private val ADDRESS = "1Maw8BjWgj6MWrBCfupqQuWANthMhefb2v"
 @RunWith(AndroidJUnit4::class)
 class BlockstackSessionTokenTest {
+
+
     @get:Rule
     val rule = ActivityTestRule(TestActivity::class.java)
 
-    private lateinit var session: BlockstackSession
+    private lateinit var blockstack: Blockstack
 
     @Before
     fun setup() {
-        session = BlockstackSession(rule.activity,
-                "https://flamboyant-darwin-d11c17.netlify.com".toBlockstackConfig(emptyArray()),
-                sessionStore = sessionStoreforIntegrationTests(rule),
-                executor = IntegrationTestExecutor(rule))
+        blockstack = Blockstack()
     }
 
-    @After
-    fun teardown() {
-        session.release()
-    }
 
     @Test
     fun extractProfileReturnsCorrectObject() {
-        val profileToken = session.extractProfile(TOKEN)
-        assertThat(profileToken.json.toString(), `is`("{\"@context\":\"http:\\/\\/schema.org\\/\",\"@type\":\"CreativeWork\",\"name\":\"Balloon Dog\",\"creator\":[{\"@type\":\"Person\",\"@id\":\"therealjeffkoons.id\",\"name\":\"Jeff Koons\"}],\"dateCreated\":\"1994-05-09T00:00:00-0400\",\"datePublished\":\"2015-12-10T14:44:26-0500\"}"))
+        val profileToken = blockstack.extractProfile(TOKEN, "")
+        assertThat(profileToken?.json.toString(), `is`("{\"@context\":\"http:\\/\\/schema.org\\/\",\"@type\":\"CreativeWork\",\"name\":\"Balloon Dog\",\"creator\":[{\"@type\":\"Person\",\"@id\":\"therealjeffkoons.id\",\"name\":\"Jeff Koons\"}],\"dateCreated\":\"1994-05-09T00:00:00-0400\",\"datePublished\":\"2015-12-10T14:44:26-0500\"}"))
     }
 
     @Test
     fun extractProfileAlsoVerifiesToken() {
-        val profileToken = session.extractProfile(TOKEN, TOKEN_PUBLIC_KEY)
-        assertThat(profileToken.json.toString(), `is`(TOKEN_PROFILE_CONTENT))
+        val profileToken = blockstack.extractProfile(TOKEN, TOKEN_PUBLIC_KEY)
+        assertThat(profileToken?.json.toString(), `is`(TOKEN_PROFILE_CONTENT))
     }
 
     @Test
     fun extractProfileThrowsForWrongOwner() {
         try {
-            session.extractProfile(TOKEN, TOKEN_PUBLIC_KEY_2)
+            blockstack.extractProfile(TOKEN, TOKEN_PUBLIC_KEY_2)
             throw AssertionError("should throw an error")
-        } catch (e: V8ScriptExecutionException) {
+        } catch (e: Exception) {
             assertThat(e.message, endsWith("Error: Token issuer public key does not match the verifying value"))
         }
     }
 
     @Test
     fun wrapProfileTokenReturnsTokenPair() {
-        val profileTokenPair = session.wrapProfileToken(TOKEN)
+        val profileTokenPair = blockstack.wrapProfileToken(TOKEN)
         assertThat(profileTokenPair.json.getJSONObject("decodedToken").toString(), `is`(DECODED_TOKEN))
         assertThat(profileTokenPair.json.getString("token"), `is`(TOKEN))
     }
 
     @Test
     fun verifyProfileTokenReturnsToken() {
-        val profileToken = session.verifyProfileToken(TOKEN, TOKEN_PUBLIC_KEY)
-        assertThat(profileToken.json.toString(), `is`(DECODED_TOKEN))
+        runBlocking {
+            val profileToken = blockstack.verifyProfileToken(TOKEN, TOKEN_PUBLIC_KEY)
+            assertThat(profileToken.json.toString(), `is`(DECODED_TOKEN))
+        }
     }
 
     @Test
     fun verifyProfileTokenThrowsForWrongOwner() {
-        try {
-            session.verifyProfileToken(TOKEN, TOKEN_PUBLIC_KEY_2)
-            throw AssertionError("should throw an error")
-        } catch (e: V8ScriptExecutionException) {
-            assertThat(e.message, endsWith("Error: Token issuer public key does not match the verifying value"))
+        runBlocking {
+            try {
+                blockstack.verifyProfileToken(TOKEN, TOKEN_PUBLIC_KEY_2)
+                throw AssertionError("should throw an error")
+            } catch (e: Exception) {
+                assertThat(e.message, endsWith("Error: Token issuer public key does not match the verifying value"))
+            }
         }
     }
 
     @Test
     fun parseZoneFileReturnsCorrectObject() {
-        val zoneFile = session.parseZoneFile(ZONE_FILE_CONTENT)
+        val zoneFile = parseZoneFile(ZONE_FILE_CONTENT)
         assertThat(zoneFile.json.toString(), `is`("{\"\$origin\":\"friedger.id\",\"\$ttl\":3600,\"uri\":[{\"name\":\"_http._tcp\",\"target\":\"https:\\/\\/gaia.blockstack.org\\/hub\\/1Maw8BjWgj6MWrBCfupqQuWANthMhefb2v\\/0\\/profile.json\",\"priority\":10,\"weight\":1}]}"))
     }
 
     @Test
     fun resolveZoneFileToProfileReturnsCorrectObject() {
-        var result: Profile? = null
-        val latch = CountDownLatch(1)
-        session.resolveZoneFileToProfile(ZONE_FILE_CONTENT) {
-            latch.countDown()
-            result = it.value
-        }
+        var result: Profile? = blockstack.resolveZoneFileToProfile(ZONE_FILE_CONTENT, ADDRESS)
 
         assertThat(result, notNullValue()) // content is dynamic
     }
 
     @Test
     fun signProfileTokenReturnsSignedToken() {
-        val tokenPair = session.signProfileToken(Profile(JSONObject(TOKEN_PROFILE_CONTENT)), PRIVATE_KEY, Entity.withKey(TOKEN_PUBLIC_KEY), Entity.withKey(TOKEN_PUBLIC_KEY_2), SigningAlgorithm.ES256K)
+        val tokenPair = runBlocking {
+            blockstack.signProfileToken(Profile(JSONObject(TOKEN_PROFILE_CONTENT)), PRIVATE_KEY, Entity.withKey(TOKEN_PUBLIC_KEY), Entity.withKey(TOKEN_PUBLIC_KEY_2))
+        }
         assertThat(tokenPair.decodedToken?.signature, notNullValue())
     }
 }
