+++ blockstack-sdktest/src/androidTest/java/org/blockstack/android/sdktest/BlockstackSession2EncryptionTest.kt
@@ -1,19 +1,21 @@
-package org.blockstack.android.sdk
+package org.blockstack.android.sdktest
 
 import android.util.Log
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.ActivityTestRule
 import okhttp3.OkHttpClient
+import org.blockstack.android.sdk.Blockstack
+import org.blockstack.android.sdk.BlockstackSession
 import org.blockstack.android.sdk.model.CryptoOptions
 import org.blockstack.android.sdk.model.toBlockstackConfig
 import org.blockstack.android.sdk.test.TestActivity
+import org.blockstack.android.sdktest.j2v8.BlockstackSessionJ2V8
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import java.util.concurrent.CountDownLatch
 
 
 private val PRIVATE_KEY = "a5c61c6ca7b3e7e55edee68566aeab22e4da26baa285c7bd10e8d2218aa3b229"
@@ -28,29 +30,29 @@ class BlockstackSession2EncryptionTest {
     @get:Rule
     val rule = ActivityTestRule(TestActivity::class.java)
 
-    private lateinit var session: BlockstackSession
+    private lateinit var sessionJ2V8: BlockstackSessionJ2V8
     private lateinit var blockstack: Blockstack
-    private lateinit var session2: BlockstackSession2
+    private lateinit var session: BlockstackSession
 
     @Before
     fun setup() {
-        val sessionStore = sessionStoreforIntegrationTests(rule)
+        val sessionStore = org.blockstack.android.sdk.sessionStoreforIntegrationTests(rule)
         val executor = IntegrationTestExecutor(rule)
         val callFactory = OkHttpClient()
-        session = BlockstackSession(rule.activity,
+        sessionJ2V8 = BlockstackSessionJ2V8(rule.activity,
                 "https://flamboyant-darwin-d11c17.netlify.com".toBlockstackConfig(emptyArray()),
                 sessionStore = sessionStore,
                 executor = executor,
                 callFactory = callFactory)
 
         // get a gaiaHubConfig by using a j2v8 call to gaia
-        session.listFiles({ false }, {
+        sessionJ2V8.listFiles({ false }, {
             val gaiaHubConfig = sessionStore.sessionData.json.getJSONObject("userData").getJSONObject("gaiaHubConfig")
             Log.d(TAG, gaiaHubConfig.toString())
         })
 
         blockstack = Blockstack()
-        session2 = BlockstackSession2(sessionStore, executor, callFactory = callFactory, blockstack = blockstack)
+        session = BlockstackSession(sessionStore, callFactory = callFactory, blockstack = blockstack)
 
         val appPrivateKey = sessionStore.sessionData.json.getJSONObject("userData").getString("appPrivateKey")
         val hubUrl = sessionStore.sessionData.json.getJSONObject("userData").getString("hubUrl")
@@ -60,7 +62,7 @@ class BlockstackSession2EncryptionTest {
     @Test
     fun testEncryptDecrypt2() {
         val message = "Hello Test"
-        val result = session.encryptContent(message, CryptoOptions(publicKey = PUBLIC_KEY))
+        val result = sessionJ2V8.encryptContent(message, CryptoOptions(publicKey = PUBLIC_KEY))
         val plainText = blockstack.decryptContent(result.value!!.json.toString(), false, CryptoOptions(privateKey = PRIVATE_KEY))
         assertThat(plainText.value as String, `is`(message))
     }
@@ -79,7 +81,7 @@ class BlockstackSession2EncryptionTest {
     fun testEncrypt2Decrypt() {
         val message = "Hello Test"
         val result = blockstack.encryptContent(message, CryptoOptions(publicKey = PUBLIC_KEY))
-        val plainText = session.decryptContent(result.value!!.json.toString(), false, CryptoOptions(privateKey = PRIVATE_KEY))
+        val plainText = sessionJ2V8.decryptContent(result.value!!.json.toString(), false, CryptoOptions(privateKey = PRIVATE_KEY))
         assertThat(plainText.value as String, `is`(message))
     }
